refactored the rhsso code as this needs in API or CLI tests
